requirements.txt
Flask-Login == 0.4.1
Flask-WTF == 0.14.3
Flask-SQLAlchemy == 2.4.4
Flask-UUID == 0.2
markupsafe == 1.1.1
psycopg2-binary == 2.8.6
pydantic == 1.6.1
six == 1.15.0
ujson == 1.35
numpy == 1.19.4
pandas == 1.1.5
scikit-learn == 0.23.2
matplotlib == 3.3.3
